requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
python-dotenv==0.19.0
psycopg2-binary==2.9.9
Werkzeug==2.0.3
SQLAlchemy==1.4.23 